Interpreter

Description

Das Modul Interpreter stellt Funktionen zur Verfügung , welche Zustände und Events interpretieren

Documentation

process :: State -> [Event] -> IO State

process ruft mit den übergebenen Parametern der Reihe nach update auf. zusätzlich kann es noch das Event Read_Output verarbeiten, welches IO Aktionen erfordert

update :: State -> Event -> State

update arbeitet ein Event in den Zustand ein und lieferten diesen Neuen Zustand zurück. folgende Event sind möglich: Add_Relation Start_Execution Trigger_Event die Funktion geht davon aus, dass das aufgerufene Event in diesem Zustand möglich ist.

update' :: Event -> State -> State

update' ist update mit verdrehter signatur, für details siehe update